+++ astromodels/functions/dark_matter/dm_models.py
@@ -6,7 +6,7 @@ import six
 from scipy.interpolate import RegularGridInterpolator
 
 from astromodels.functions.function import Function1D, FunctionMeta
-from astromodels.utils.data_files import _get_data_file_path
+from astromodels.utils import _get_data_file_path
 from astromodels.utils.logging import setup_logger
 
 log = setup_logger(__name__)
