+++ astromodels/tests/test_1D_function_values.py
@@ -6,7 +6,7 @@ import h5py
 
 from astromodels.functions.priors import *
 from astromodels.functions.function import _known_functions
-from astromodels.utils.data_files import _get_data_file_path
+from astromodels.utils import _get_data_file_path
 
 _multiplicative_models = ["PhAbs", "TbAbs", "WAbs", "APEC", "VAPEC", "EBLattenuation" ]
