+++ astromodels/utils/__init__.py
@@ -1,1 +1,3 @@
 __author__ = 'giacomov'
+
+from .file_utils import _get_data_file_path, get_user_data_path, get_user_path
